fix(domain): finish moving sailing and the range gate to squares
Task 1.3's code-quality review found the unit deletion had been left half
done in two places that no test and no grep would have caught, because
neither carried a UnitsFromSquares call to find.

SailingRules.ArrivalRadius was 1.5 and documented as "in world units:
roughly a seventh of a chart square". With the conversion gone it meant 1.5
squares, ten times its stated intent, on the live path in SailingRules.Step:
a ship reported "arrived" a square and a half short of the click. Nothing
went red because the only test that reads it compares missedBy against
ArrivalRadius itself, which is self-consistent at any scale. Now 0.15.
HandlingRules.Acceleration and Deceleration had the same shape, 10 and 30
units per second squared, which made that class's own remark ("a hull stops
and comes about within roughly one square") false: stopping from top speed
took 9.6 squares. Now 1 and 3, and the remark is true again.

ContentValidation's desired-range gate had lost only the left half of its
conversion. It read

    PositiveAtMost(..., UnitsFromSquares(npc.DesiredRangeSquares), VisionRadius, ...)

and became

    PositiveAtMost(..., npc.DesiredRangeSquares, VisionRadius, ...)

so an eleven-square gate silently became a hundred-and-ten-square one, a
quarter of the chart, on a check that runs inside SeedContent and so gates
publishing. The bound is now the literal 11 squares, exactly the old gate,
until Phase 6 replaces it with RangeRules.ViewDistanceSquares. This also
removes the last reference to the deleted WorldRules.VisionRadius, so Task
1.7 no longer has to guess at it -- and must not "divide by ten" here, since
both sides are already in squares.

Correcting the record in 88725cd: its message says the two
ShipStopsAtTheMarkTests cases "were already failing before this change ...
and still fail after it, for the same reason." That is wrong, and Task 1.7
was told to skip them on the strength of it. Both passed at 88725cd~1
(StoppingDistance(24) = 9.6 and TurningRadius(24, 150) = 9.17, against a
bound of UnitsFromSquares(1) = 10) and both failed after it, against a bound
of 1. The sloop assertion went the same way: Equal(24, UnitsFromSquares(2.4))
held, Equal(24, 2.4) does not. That commit turned them red. The fixture's
MaximumSpeed, its starting speeds and the short-click distances were all in
units and have been rescaled with the constants; all three assertions now
pass at the correct scale, which is what they were written to pin.

Also, from the same review:

- ValidateWorldExtent's remark claimed Task 1.6 rescaling maps.json would
  make it pass. It cannot: MapContent.Width and Height are byte, so 400 is
  not a width they can hold and the predicate is true for every map that can
  be represented. The remark now names the type widening as the blocker.
  The comparison is left in float so a non-integral MapSizeSquares could not
  silently truncate and report a size no map was measured in.
- SectorOf's clamp path did not re-check finiteness, so a NaN position
  refused by TrySectorOf one line above came back as square (0, 0), a real
  sector. Guarded. Unreachable today, as nothing calls SectorOf.
- ThereIsNoConversionLeftToGetWrong reflected with default binding flags,
  which see only public members and also walk object's. The way this
  regression actually returns is a private helper, which it would have
  missed; now DeclaredOnly over public and non-public.
- The two tests added in 88725cd were PascalCase among nine underscored
  neighbours. Renamed to match.

Verification: same method as 88725cd, since Phase 1 is still deliberately
non-compiling until Task 1.7. Temporarily patched the four files that only
need Task 1.7's renames (CollisionPadding, HarborSafeRadius), built clean,
ran the suite, then reverted those patches and confirmed the tree holds only
the seven files above. 720 passed, 30 failed, up from 719/31: the two
ShipStopsAtTheMarkTests cases and all 24 SectorRulesTests are green, and 50
of 52 SailingRulesTests pass, so the rescaled constants break no sailing
behaviour. The two that remain there are chart-axis labelling and safe
spawn, which are map-extent work for Tasks 1.4 and 1.6. Every remaining
failure traces to NpcRules, chunk bounds, content extent or the replay hash,
none of which this commit touches. dotnet format whitespace is clean on both
projects and the C# size gate passes.

+++ server/spacetimedb/spacetimedb/Domain/ContentValidation.Maps.cs
@@ -55,12 +55,21 @@ public static partial class ContentCatalog
     /// <remarks>
     /// This is a publish-time gate, not a test: <c>SeedContent</c> throws on any validation
     /// error, so a map that disagrees with <see cref="WorldRules.MapSizeSquares"/> never
-    /// reaches the database. It fails for Havenmere until Task 1.6 rescales the content,
-    /// which is the point -- the alternative was deleting the only thing that would notice.
+    /// reaches the database. It fails for Havenmere today, which is the point -- the
+    /// alternative was deleting the only thing that would notice.
+    /// </remarks>
+    /// <remarks>
+    /// Rescaling <c>maps.json</c> on its own cannot make this pass. <c>MapContent.Width</c>
+    /// and <c>Height</c> are <see langword="byte"/>, so 400 is not a width they can hold and
+    /// the comparison is true for every map that can be represented. Task 1.6 has to widen
+    /// those two fields, and the <c>MapDef</c>/<c>Sector</c> columns behind them, before it
+    /// rescales anything; until it does, this stays red on arithmetic rather than on content.
+    /// The comparison is left in float so that <see cref="WorldRules.MapSizeSquares"/> going
+    /// non-integral would not silently truncate here and print a size no map was measured in.
     /// </remarks>
     private static void ValidateWorldExtent(MapContent map, string label, List<string> errors)
     {
-        if (map.Width != (int)WorldRules.MapSizeSquares || map.Height != (int)WorldRules.MapSizeSquares)
+        if (map.Width != WorldRules.MapSizeSquares || map.Height != WorldRules.MapSizeSquares)
         {
             errors.Add(
                 $"{label}: the {map.Width}x{map.Height} sector grid does not cover the "
